src/requests/requests.controller.ts
import { Body, Controller, HttpStatus, Post, Res } from '@nestjs/common';
import { Response } from 'express';
import { CreateRequestDto } from './dto/create-request.dto';
import { RequestsService } from './requests.service';

@Controller('requests')
export class RequestsController {
  constructor(private readonly requestsService: RequestsService) {}

  @Post()
  async create(
    @Body() createRequestDto: CreateRequestDto,
    @Res() res: Response,
  ) {
    await this.requestsService.create(createRequestDto);
    res.status(HttpStatus.CREATED).send();
  }
}
